Refactor the register function.
- modularise
- add test file
- made validations module
