+++ lib/index.js
@@ -1,23 +1,12 @@
-var version = require('./version');
-var init = require('./init');
-var test = require('./test');
-var publish = require('./publish');
-var register = require('./register');
-var exportResume = require('./exportResume');
-var settings = require('./settings');
-var serve = require('./serve');
-var login = require('./login');
-var waterfallArray = require('./waterfallArray');
-
 module.exports = {
-    init: init,
-    test: test,
-    publish: publish,
-    register: register,
-    exportResume: exportResume,
-    settings: settings,
-    serve: serve,
-    version: version,
-    login: login,
-    waterfallArray: waterfallArray
+    init: require('./init'),
+    test: require('./test'),
+    publish: require('./publish'),
+    register: require('./register'),
+    exportResume: require('./exportResume'),
+    settings: require('./settings'),
+    serve: require('./serve'),
+    version: require('./version'),
+    login: require('./login'),
+    waterfallArray: require('./waterfallArray')
 };
