--- test/index.js
@@ -1,8 +1,0 @@
-var lib = require('../lib');
-var test = require('tape');
-var fs = require('fs');
-
-test('test.', function(t) {
-    t.plan(1);
-    t.ok(true, 'todo: write tests....');
-});
