+++ test/register-test.js
@@ -1,0 +1,14 @@
+var should = require('should');
+
+describe("Register tests", function() {
+//
+	it('should register new user', function(done) {
+//
+//
+//
+//
+		done();
+	});
+//
+});
+console.log('jeje');
